+++ example/lib/main.dart
@@ -39,41 +39,48 @@ class _HomePageState extends State<HomePage> {
           children: [
             ElevatedButton(
               onPressed: () async {
-                var res = await Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const SimpleBarcodeScannerPage(
-                        barcodeAppBar: BarcodeAppBar(
-                          appBarTitle: 'Test',
-                          centerTitle: false,
-                          enableBackButton: true,
-                          backButtonIcon: Icon(Icons.arrow_back_ios),
-                        ),
-                        delayMillis: 2000,
-                        child: Column(
-                          children: [
-                            SizedBox(
-                              height: 20,
-                            ),
-                            TextField(
-                              decoration: InputDecoration(
-                                labelText: 'Enter Text',
-                                border: OutlineInputBorder(),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ));
+                String? res = await SimpleBarcodeScanner.scanBarcode(
+                  context,
+                  barcodeAppBar: const BarcodeAppBar(
+                    appBarTitle: 'Test',
+                    centerTitle: false,
+                    enableBackButton: true,
+                    backButtonIcon: Icon(Icons.arrow_back_ios),
+                  ),
+                  isShowFlashIcon: true,
+                  delayMillis: 2000,
+                );
                 setState(() {
-                  if (res is String) {
-                    result = res;
-                  }
+                  result = res as String;
                 });
               },
-              child: const Text('Open Scanner'),
+              child: const Text('Scan Barcode'),
             ),
-            Text('Barcode Result: $result'),
+            const SizedBox(
+              height: 10,
+            ),
+            Text('Scan Barcode Result: $result'),
+            const SizedBox(
+              height: 10,
+            ),
+            ElevatedButton(
+              onPressed: () async {
+                SimpleBarcodeScanner.streamBarcode(
+                  context,
+                  barcodeAppBar: const BarcodeAppBar(
+                    appBarTitle: 'Test',
+                    centerTitle: false,
+                    enableBackButton: true,
+                    backButtonIcon: Icon(Icons.arrow_back_ios),
+                  ),
+                  isShowFlashIcon: true,
+                  delayMillis: 2000,
+                ).listen((event) {
+                  print("Stream Barcode Result: $event");
+                });
+              },
+              child: const Text('Stream Barcode'),
+            ),
           ],
         ),
       ),