+++ lib/constant.dart
@@ -6,3 +6,4 @@ class PackageConstant {
 }
 
 String kScanPageTitle = 'Scan barcode/qrcode';
+String kCancelValue = '-2';