+++ lib/screens/io_device.dart
@@ -5,6 +5,7 @@ import 'package:simple_barcode_scanner/enum.dart';
 import 'package:simple_barcode_scanner/screens/window.dart';
 
 import '../barcode_appbar.dart';
+import '../constant.dart';
 import '../flutter_barcode_scanner.dart';
 
 /// Barcode scanner for mobile and desktop devices
@@ -19,20 +20,20 @@ class BarcodeScanner extends StatelessWidget {
   final Widget? child;
   final BarcodeAppBar? barcodeAppBar;
   final int? delayMillis;
-
-  const BarcodeScanner({
-    super.key,
-    required this.lineColor,
-    required this.cancelButtonText,
-    required this.isShowFlashIcon,
-    required this.scanType,
-    required this.onScanned,
-    this.child,
-    this.appBarTitle,
-    this.centerTitle,
-    this.barcodeAppBar,
-    this.delayMillis,
-  });
+  final Function? onClose;
+  const BarcodeScanner(
+      {super.key,
+      required this.lineColor,
+      required this.cancelButtonText,
+      required this.isShowFlashIcon,
+      required this.scanType,
+      required this.onScanned,
+      this.child,
+      this.appBarTitle,
+      this.centerTitle,
+      this.barcodeAppBar,
+      this.delayMillis,
+      this.onClose});
 
   @override
   Widget build(BuildContext context) {
@@ -50,7 +51,24 @@ class BarcodeScanner extends StatelessWidget {
       );
     } else {
       /// Scan Android and ios barcode scanner with flutter_barcode_scanner
-      _scanBarcodeForMobileAndTabDevices();
+      /// If onClose is not null then stream barcode otherwise scan barcode
+      /// Scan barcode for mobile devices
+      ScanMode scanMode;
+      switch (scanType) {
+        case ScanType.barcode:
+          scanMode = ScanMode.BARCODE;
+          break;
+        case ScanType.qr:
+          scanMode = ScanMode.QR;
+          break;
+        default:
+          scanMode = ScanMode.DEFAULT;
+          break;
+      }
+      onClose != null
+          ? _streamBarcodeForMobileAndTabDevices(scanMode)
+          : _scanBarcodeForMobileAndTabDevices(scanMode);
+
       return const Scaffold(
         body: Center(
           child: CircularProgressIndicator(),
@@ -59,22 +77,19 @@ class BarcodeScanner extends StatelessWidget {
     }
   }
 
-  _scanBarcodeForMobileAndTabDevices() async {
-    /// Scan barcode for mobile devices
-    ScanMode scanMode;
-    switch (scanType) {
-      case ScanType.barcode:
-        scanMode = ScanMode.BARCODE;
-        break;
-      case ScanType.qr:
-        scanMode = ScanMode.QR;
-        break;
-      default:
-        scanMode = ScanMode.DEFAULT;
-        break;
-    }
+  _scanBarcodeForMobileAndTabDevices(ScanMode scanMode) async {
     String barcode = await FlutterBarcodeScanner.scanBarcode(
         lineColor, cancelButtonText, isShowFlashIcon, scanMode, delayMillis);
     onScanned(barcode);
+  }
+
+  void _streamBarcodeForMobileAndTabDevices(ScanMode scanMode) {
+    FlutterBarcodeScanner.getBarcodeStreamReceiver(
+            lineColor, cancelButtonText, isShowFlashIcon, scanMode, delayMillis)
+        ?.listen((barcode) {
+      if (barcode != null) {
+        barcode == kCancelValue ? onClose?.call() : onScanned(barcode);
+      }
+    });
   }
 }