+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.9
+* `SimpleBarcodeScannerPage` is now depreciated, use `SimpleBarcodeScanner` instead
+* Scan barcode with `SimpleBarcodeScanner` `scanBarcode` method
+* Stream barcode with `SimpleBarcodeScanner` `streamBarcode` method
+
 ## 0.1.8
 * IOS build issue fixed
 * Delay feature added for android and ios